build(nrf52): -Wall -Wextra -Werror fuer wiscore_rak4631 (Wave 0.2 Rest)
Wave 0.2 hatte -Werror bereits fuer die 23 ESP32-Envs aktiviert
(deba2ad8), die drei nRF52-Boards blieben offen. Fuer wiscore_rak4631
(angeschlossene Hardware) jetzt nachgezogen.

Vor dem Gate 20 echte Warnungen in src/ gemessen (nicht die ~33.000
aus SVCALL-Makro-Expansionen in den Nordic-SoftDevice-Headern -- die
sind aus Drittanbieter-Headern und via -Wno-unused-parameter in
build_src_flags gezielt ausgeklammert, analog zum bestehenden
-Wno-missing-field-initializers-Muster fuer ESP32). Alle 20 behoben,
durchweg (void)-Cast auf tatsaechlich ungenutzte Parameter bzw.
int -> size_t bei zwei sizeof-Vergleichsschleifen in nrf52_main.cpp:

- Regexp.cpp, aht20.cpp, bmp390.cpp, aprs_functions.cpp (x2),
  extudp_functions.cpp, lora_functions.cpp, loop_functions.cpp,
  nrf52/WisBlock-API.cpp, nrf52/nrf52_ble.cpp: ungenutzte Parameter
  markiert, keine Verhaltensaenderung.
- lora_setchip.cpp: die sieben Parameter von lora_setchip_new() sind
  nur im ESP32/RadioLib-Zweig genutzt; auf jeder anderen Plattform
  (inkl. nRF52, das SX126x-Arduino statt RadioLib nutzt) kompiliert
  die Funktion zum No-op. (void)-Casts jetzt unbedingt am Anfang,
  wirkungslos auf ESP32 (Parameter werden dort weiterhin normal
  benutzt).
- nrf52_main.cpp: `for (int i = 0; i < sizeof(...); i++)` in
  sendUBX_MON_VER/sendUBX_SET_GNSS auf size_t umgestellt
  (-Wsign-compare).

heltec_t114/t_echo bewusst NICHT mit angefasst: das Aktivieren von
-Wall/-Werror dort deckte einen davon unabhaengigen, groesseren
Befund auf (Sektion `[nrf52]` kollidiert namensgleich ueber alle drei
variants/*/platformio.ini, siehe naechster Commit) -- ohne
angeschlossene Hardware fuer diese zwei Boards nicht im
Zero-Tolerance-Rahmen verifizierbar. Als eigener Folgepunkt
zurueckgestellt.

Verifikation: wiscore_rak4631 sauber gebaut, RAM/Flash unveraendert
(81988 B / 576016 B). Voller Fleet-Build (`pio run`, alle default_envs)
sowie `pio test -e native` zusaetzlich gegengeprueft.

Refs: Wave 0.2 Rest (resume.md, docs/architecture/08-defect-catalogue.md)

+++ src/nrf52/nrf52_main.cpp
@@ -2549,7 +2549,7 @@ void sendUBX_MON_VER() {  // Binäres Paket senden
   if(iGPSDEBUG >= 2)
     Serial.println("[GPS ]...Sende UBX_MON_VER");
 
-  for (int i = 0; i < sizeof(UBX_MON_VER_RAK); i++)
+  for (size_t i = 0; i < sizeof(UBX_MON_VER_RAK); i++)
   {
     Serial1.write(UBX_MON_VER_RAK[i]);
   }
@@ -2562,7 +2562,7 @@ void sendUBX_SET_GNSS() {  // Binäres Paket senden
   if(iGPSDEBUG >= 2)
     Serial.println("[GPS ]...Sende UBX_SET_GNSS");
 
-  for (int i = 0; i < sizeof(ubx_cfg_gnss); i++)
+  for (size_t i = 0; i < sizeof(ubx_cfg_gnss); i++)
   {
     Serial1.write(ubx_cfg_gnss[i]);
   }
